WIP adding sponsor management system

+++ routes/admin.php
@@ -27,12 +27,13 @@ Route::delete('/group-management/{groupID}', 'admin\GroupManagementController@de
 Route::patch('/group-management/{groupID}', 'admin\GroupManagementController@restore')->name('admin.groupManagement.restore');
 
 //View, create delete, assign sponsors
-Route::get('/sponsor-management', 'SponsorManagementController@index')->name('admin.sponsorManagement');
-Route::get('/sponsor-management/{sponsorID}', 'SponsorManagementController@view')->name('admin.sponsorManagement.view');
-Route::post('/sponsor-management', 'SponsorManagementController@store')->name('admin.sponsorManagement.store');
-Route::put('/sponsor-management/{sponsorID}/{eventID}', 'SponsorManagementController@addToEvent')->name('admin.sponsorManagement.addToEvent');
-Route::delete('/sponsor-management/{SponsorID}', 'SponsorManagementController@deleteSponsor')->name('admin.sponsorManagement.deleteSponsor');
-Route::delete('/sponsor-management/{SponsorID}/{eventID}', 'SponsorManagementController@removeFromEvent')->name('admin.sponsorManagement.removeFromEvent');
+Route::get('/sponsor-management', 'admin\SponsorManagementController@index')->name('admin.sponsorManagement');
+Route::get('/sponsor-management/{sponsorID}', 'admin\SponsorManagementController@view')->name('admin.sponsorManagement.view');
+Route::post('/sponsor-management', 'admin\SponsorManagementController@store')->name('admin.sponsorManagement.store');
+Route::patch('/sponsor-management/{sponsorID}', 'admin\SponsorManagementController@update')->name('admin.sponsorManagement.update');
+Route::put('/sponsor-management/{sponsorID}/{eventID}', 'admin\SponsorManagementController@addToEvent')->name('admin.sponsorManagement.addToEvent');
+Route::delete('/sponsor-management/{SponsorID}', 'admin\SponsorManagementController@deleteSponsor')->name('admin.sponsorManagement.deleteSponsor');
+Route::delete('/sponsor-management/{SponsorID}/{eventID}', 'admin\SponsorManagementController@removeFromEvent')->name('admin.sponsorManagement.removeFromEvent');
 
 //View, Update Group
 Route::get('/group-management/{groupID}', 'admin\GroupManagementController@view')->name('admin.group');
